fix directory names
