+++ tests/FactroApiClient.UnitTests/CompanyApi/CompanyApiTests.cs
@@ -1,4 +1,4 @@
-namespace FactroApiClient.UnitTests.Company
+namespace FactroApiClient.UnitTests.CompanyApi
 {
     using System.Diagnostics.CodeAnalysis;
